requirements.txt
# Core dependencies
streamlit>=1.28.0
groq>=0.3.0
beautifulsoup4>=4.12.0
selenium>=4.15.0
requests>=2.31.0
pandas>=2.0.0
numpy>=1.24.0
plotly>=5.18.0

# Database
sqlalchemy>=2.0.0

# Utilities
python-dotenv>=1.0.0
fake-useragent>=1.4.0
retry>=0.9.2
tqdm>=4.66.0
webdriver-manager>=4.0.0

# Data processing
nltk>=3.8.0
textblob>=0.17.0
wordcloud>=1.9.0

# Development
pytest>=7.4.0
black>=23.0.0
flake8>=6.1.0
